auth/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.7
PyJWT==2.10.1
Werkzeug==3.1.3
